core: remove dead code, limit test code scope (#17006)
* core: move test util var/func to test file

* core: remove useless func

* core: remove String methods
